refactor: move markdown internal transformations to printer.preprocess
