+++ src/language-markdown/utils.js
@@ -47,7 +47,23 @@ function getFencedCodeBlockValue(node, originalText) {
   }
 }
 
+function mapAst(ast, handler) {
+  return (function preorder(node, index, parentStack) {
+    parentStack = parentStack || [];
+
+    const newNode = Object.assign({}, handler(node, index, parentStack));
+    if (newNode.children) {
+      newNode.children = newNode.children.map((child, index) => {
+        return preorder(child, index, [newNode].concat(parentStack));
+      });
+    }
+
+    return newNode;
+  })(ast, null, null);
+}
+
 module.exports = {
+  mapAst,
   getFencedCodeBlockValue,
   getOrderedListItemInfo
 };
